feat: add artefact_selection_schema module and define stage schemas
