+++ bitloops/src/graphql/types.rs
@@ -1,5 +1,6 @@
 pub mod artefact;
 pub mod artefact_selection;
+mod artefact_selection_schema;
 pub mod chat;
 pub mod checkpoint;
 pub mod clone;
